Converted integer value to decimal value

+++ src/Model/Table/BalanceReceiverDetailsTable.php
@@ -76,19 +76,19 @@ class BalanceReceiverDetailsTable extends Table
             ->notEmptyString('transaction');
 
         $validator
-            ->integer('payable')
+            ->decimal('payable')
             ->allowEmptyString('payable');
 
         $validator
-            ->integer('receivable')
+            ->decimal('receivable')
             ->allowEmptyString('receivable');
 
         $validator
-            ->integer('received')
+            ->decimal('received')
             ->allowEmptyString('received');
 
         $validator
-            ->integer('sent')
+            ->decimal('sent')
             ->allowEmptyString('sent');
 
         $validator
